# apps/nest-backend/openapi/admin.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/providers": {
      "get": {
        "operationId": "adminAuthProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthProvidersDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос или неизвестные входные поля.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учётные данные или токен либо неактивная сессия.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Нет прав на действие либо некорректные заголовки браузерной авторизации.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден или недоступен текущему пользователю.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт с существующими данными или повторное действие.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "413": {
            "description": "Превышен лимит: 64 КиБ для JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "415": {
            "description": "Неподдерживаемое кодирование содержимого или кодировка символов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Превышен лимит запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Получить доступные способы входа администратора",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/keycloak/complete": {
      "post": {
        "description": "Без тела. Обменять browser-bound completion cookie на собственную административную сессию и Bearer JWT в accessToken на 7 дней (604800 секунд). Cookie авторизации не устанавливается; временные OIDC cookies очищаются. Вызывать один раз под межвкладочной блокировкой; после потери ответа начать новый вход, не повторять запрос.",
        "operationId": "adminKeycloakComplete",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "description": "Корректный HTTP(S) Origin без пути, query или fragment. Принимается любой источник; allowlist не используется.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Protection",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос или неизвестные входные поля.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Completion отсутствует, просрочен, использован или доступ отозван.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Обязательны корректный HTTP(S) Origin и X-CSRF-Protection: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Keycloak отключён.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт с существующими данными или повторное действие.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "413": {
            "description": "Превышен лимит: 64 КиБ для JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "415": {
            "description": "Неподдерживаемое кодирование содержимого или кодировка символов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-keycloak-completion": []
          },
          {
            "admin-keycloak-completion-secure": []
          }
        ],
        "summary": "Однократно завершить вход через Keycloak",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Starter — административное API",
    "description": "Административное API с настраиваемыми ролями. Первый владелец создаётся при запуске на пустой БД; дополнительные аккаунты — через OWNER API. Роль и permissions доступны в /auth/me; требования операции — в x-admin-permissions. OWNER имеет все права, ADMIN — все несистемные, USER — собственный аккаунт. Управление доступом и Keycloak доступно только OWNER.\n\nВход и выход через браузер требуют заголовков Origin и X-CSRF-Protection: 1. Принимается любой корректный HTTP(S) Origin без allowlist; CORS открыт.\n\nАвторизация — только Bearer JWT в заголовке Authorization. При входе JWT возвращается в accessToken, cookie авторизации не устанавливаются. JWT и сессия действуют 7 дней (604800 секунд), включая Keycloak. Продления нет; после истечения срока нужен новый вход.\n\nLogout требует валидный JWT текущей сессии и немедленно отзывает её; повторный запрос возвращает 401. Успешная смена credentials отзывает все сессии аккаунта. Клиент удаляет сохранённый JWT после выхода, успешной смены credentials или терминального отказа Bearer. Исключение: 401 CURRENT_PASSWORD_INVALID при неверном текущем пароле не отзывает сессию и не требует удаления JWT. Клиентские и административные JWT не взаимозаменяемы.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin-keycloak-completion": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_oidc_complete",
        "description": "Одноразовый browser-bound grant после OIDC callback; дополнительно требуется browser-binding cookie. Только для завершения входа, не для авторизации API."
      },
      "admin-keycloak-completion-secure": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-admin_oidc_complete",
        "description": "HTTPS-вариант одноразового completion grant; имя и Secure определяются callback URL из настроек Keycloak в БД."
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "AccessTokenDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Bearer JWT для Authorization. Выдаётся только в JSON, не в cookie."
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "number",
            "description": "Фиксированный срок JWT: exp - iat = 604800 секунд (7 дней), включая вход через Keycloak. По истечении срока требуется новый вход.",
            "enum": [
              604800
            ],
            "example": 604800
          },
          "sessionExpiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Фиксированный срок окончания сессии, совпадающий с exp JWT. Сессия может быть отозвана раньше при выходе, смене credentials или отзыве доступа."
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn",
          "sessionExpiresAt"
        ]
      },
...
      "AdminAuthProvidersDto": {
        "type": "object",
        "properties": {
          "local": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "keycloak": {
            "type": "boolean",
            "description": "Keycloak включён и сконфигурирован. Это не проверка доступности провайдера."
          }
        },
        "required": [
          "local",
          "keycloak"
        ]
      },
...
    }
  }
}